29: Create eyebrow component and include in title. Making for reuse in a page design pattern
